+++ CHANGELOG.md
@@ -6,6 +6,11 @@
 
 - Fixed an issue for Vue.defineField method.
 
+### v1.0.4
+
+- Add prototype.sleep method.
+- Add prototype.wait method.
+
 <br />
 <br />
 <br />